+++ prereqs/install_dkinst_pipx_debian.sh
@@ -9,6 +9,9 @@ source ~/.bashrc
 echo "Installing dkinst via pipx..."
 pipx install dkinst
 
+source ~/.bashrc
+dkinst prereqs
+
 echo "dkinst installation complete. You can now run 'dkinst' from the command line."
 echo "To update dkinst in the future, run: pipx upgrade dkinst"
 echo "To uninstall dkinst, run: pipx uninstall dkinst"